feat(DiscountType): DiscountEvent, 이벤트 이름명을 갖춘 DiscountType 생성
