+++ docs/README.md
@@ -63,9 +63,9 @@
 
 ### 헤택 금액에 따른 12월 이벤트 배지 부여
 혜택 금액에 따라 서로 다른 이벤트 배지를 부여한다
-- 5천원 이상: 별
-- 1만원 이상: 트리
-- 2만원 이상: 산타
+- [x] 5천원 이상: 별
+- [x] 1만원 이상: 트리
+- [x] 2만원 이상: 산타
 
 ### 12월 이벤트 관련 유의사항
 - 이벤트 적용 가능한 금액은 총 주문 금액을 기준으로 1만원 이상이다
@@ -159,3 +159,6 @@
   - 5천원 이상: 별
   - 1만원 이상: 트리
   - 2만원 이상: 산타
+
+Day를 넘겨주면 할인 적용 할 수 있는 이벤트를 List<>로 반환한다 
+
