--- src/main/java/christmas/discount/DiscountEventName.java
@@ -1,9 +1,0 @@
-package christmas.discount;
-
-public class DiscountEventName {
-    private final String eventNameVaue;
-
-    public DiscountEventName(String eventNameVaue) {
-        this.eventNameVaue = eventNameVaue;
-    }
-}
