+++ src/main/java/christmas/domain/menu/MenuPrice.java
@@ -1,4 +1,4 @@
-package christmas.menu;
+package christmas.domain.menu;
 
 import java.util.Objects;
